+++ Core/Casa.Application/Models/TRCPartRequestModel.cs
@@ -57,9 +57,9 @@ namespace CLN.Application.Models
 
         public int? TypeOfBMSId { get; set; }
 
-        public int? AvailableQty { get; set; }
+        public decimal? AvailableQty { get; set; }
 
-        public int? RequiredQty { get; set; }
+        public decimal? RequiredQty { get; set; }
 
         [DefaultValue("")]
         public string? Remarks { get; set; }
@@ -110,8 +110,8 @@ namespace CLN.Application.Models
         public string UOMName { get; set; }
         public int? TypeOfBMSId { get; set; }
         public string TypeOfBMS { get; set; }
-        public int? AvailableQty { get; set; }
-        public int? RequiredQty { get; set; }
+        public decimal? AvailableQty { get; set; }
+        public decimal? RequiredQty { get; set; }
         public string Remarks { get; set; }
         public bool? RGP { get; set; }
     }
